+++ app/src/main/java/com/example/triviaclient/communicator/ResponseCode.java
@@ -4,9 +4,9 @@ public enum ResponseCode {
     ERROR(0),
     SUCCESS(1);
 
-    final int value;
+    final byte value;
 
     ResponseCode(int value) {
-        this.value = value;
+        this.value = (byte) value;
     }
 }
